// src/life-record/life-record-error-code.enum.ts
export enum LifeRecordErrorCode {
  INVALID_DATE_FORMAT = 'INVALID_DATE_FORMAT',
  INVALID_LIFE_VALUE = 'INVALID_LIFE_VALUE',
  INVALID_FOOD_ID = 'INVALID_FOOD_ID',
  INVALID_MEDICINE_ID = 'INVALID_MEDICINE_ID',
  LIFE_RECORD_ALREADY_EXISTS = 'LIFE_RECORD_ALREADY_EXISTS',
  LIFE_RECORD_CREATE_FAILED = 'LIFE_RECORD_CREATE_FAILED',
  LIFE_RECORD_FORBIDDEN = 'LIFE_RECORD_FORBIDDEN',
  LIFE_RECORD_NOT_FOUND = 'LIFE_RECORD_NOT_FOUND',
  TEXT_REQUIRED = 'TEXT_REQUIRED',
  TEXT_TOO_LONG = 'TEXT_TOO_LONG',
  TAG_EXTRACTION_FAILED = 'TAG_EXTRACTION_FAILED',
}